CRUD_JWT/requirements.txt
asgiref==3.5.2
Django==3.2.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
install==1.3.5
PyJWT==1.7.1
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
